Fix #30 Add a check on distance since last update
This commit will add a check to avoid updates if the camera has not
moved a certain distance since last time there was an update. Adds a new
CBA setting to allow the user to configure how large this distance
should be. This will also change the other settings to force an update
if they were changed.